add more options

+++ index.js
@@ -35,6 +35,10 @@ AliyunossWebpackPlugin.prototype.apply = function(compiler) {
 
 AliyunossWebpackPlugin.prototype.oposs = function() {
 	var _this = this;
+  var deleteAll = _this.options.deleteAll || false;
+  var generateObjectPath = _this.options.generateObjectPath || function(fileName) { return fileName; }
+  var getObjectHeaders = _this.options.getObjectHeaders || function() { return {}; }
+
 	co(function*() {
 		'use strict';
 		var store = oss({
@@ -43,27 +47,40 @@ AliyunossWebpackPlugin.prototype.oposs = function() {
 			accessKeySecret: _this.options.accessKeySecret,
 			bucket: _this.options.bucket
 		});
+
 		//删除oss上代码
-		var fileList = yield store.list();
-		var files = [];
-		if (fileList.objects) {
-			fileList.objects.forEach(function(file) {
-				files.push(file.name);
-			})
-			var result = yield store.deleteMulti(files, {
-				quiet: true
-			});
-		}
+    if (deleteAll) {
+      var fileList = yield store.list();
+      var files = [];
+      if (fileList.objects) {
+        fileList.objects.forEach(function(file) {
+          files.push(file.name);
+        })
+        var result = yield store.deleteMulti(files, {
+          quiet: true
+        });
+      }
+    }
+
 		//上传oss的新代码
 		fileUtils.eachFileSync(_this.options.buildPath, function(filename, stats) {
 			_this.fileArray.push(filename);
 		});
+
 		var j = 0;
 		for (var i = 0; i < _this.fileArray.length; i++) {
 			var file = _this.fileArray[i];
-			var fileName = file.split('/').pop();
-			yield store.put(fileName, file);
-			console.log(file + ' -- upload success'.green);
+			var fileName = file.replace(_this.options.buildPath, '');
+      if (fileName.startsWith('/')) {
+        fileName = fileName.substr(1);
+      }
+      var ossFileName = generateObjectPath(fileName);
+      if (ossFileName) {
+        yield store.put(ossFileName, file, { headers: getObjectHeaders(fileName) });
+        console.log(file + ' -- upload to ' + ossFileName + ' success'.green);
+      } else {
+        console.log('skipping file ' + file)
+      }
 		}
 	}).catch(function(err) {
 		console.info(err)
